+++ src/components/McuUpdateDialogView.tsx
@@ -13,13 +13,18 @@ import {
     classNames,
     clearConfirmBeforeClose,
     clearWaitForDevice,
+    convertToDropDownItems,
     DialogButton,
+    Dropdown,
+    type DropdownItem,
     GenericDialog,
     getPersistentStore,
+    getSelectedDropdownItem,
     logger,
     NumberInlineInput,
     Overlay,
     selectedDevice,
+    selectedDeviceInfo,
     setWaitForDevice,
     Slider,
     Toggle,
@@ -53,8 +58,19 @@ const McuUpdateDialogView = () => {
     const isVisible = useSelector(getShowMcuBootProgrammingDialog);
     const mcubootFwPath = useSelector(getMcubootFilePath);
     const zipFilePath = useSelector(getZipFilePath);
+    const deviceInfo = useSelector(selectedDeviceInfo);
+
+    const programmingOptions =
+        deviceInfo?.mcuStateOptions?.filter(s => s.type === 'Programming') ??
+        [];
+
+    const targetDropdownItems = convertToDropDownItems(
+        programmingOptions.map(s => s.arguments?.target),
+        false,
+    );
 
     const fwPath = mcubootFwPath || zipFilePath;
+    const [chosenTarget, setChosenTarget] = useState<string>('');
 
     const writingHasStarted = writing || writingFail || writingSucceed;
 
@@ -73,6 +89,9 @@ const McuUpdateDialogView = () => {
         autoStart: false,
         resolution: 200,
     });
+
+    const mustHaveChipTarget = programmingOptions.length > 1 && !zipFilePath;
+    const hasChipTarget = mustHaveChipTarget && !chosenTarget;
 
     useEffect(() => {
         // note: check may be redundant as Thingy:91 has a different modal
@@ -108,6 +127,12 @@ const McuUpdateDialogView = () => {
         }
     }, [device, showDelayTimeout]);
 
+    useEffect(() => {
+        if (targetDropdownItems.length > 0 && !chosenTarget) {
+            setChosenTarget(targetDropdownItems[0].value);
+        }
+    }, [chosenTarget, targetDropdownItems]);
+
     const onCancel = () => {
         dispatch(clearWaitForDevice());
         dispatch(setShowMcuBootProgrammingDialog(false));
@@ -126,6 +151,11 @@ const McuUpdateDialogView = () => {
 
         if (!fwPath) {
             logger.error('No file selected');
+            return;
+        }
+
+        if (hasChipTarget) {
+            logger.error('No target selected');
             return;
         }
 
@@ -172,9 +202,10 @@ Are you sure you want to continue?`,
                 setProgress(updatedProgress);
             },
             abortController.current,
-            showDelayTimeout
+            showDelayTimeout || hasChipTarget
                 ? {
                       netCoreUploadDelay: uploadDelay,
+                      target: mcubootFwPath ? chosenTarget : undefined,
                   }
                 : undefined,
         )
@@ -232,7 +263,12 @@ Are you sure you want to continue?`,
                     <DialogButton
                         variant="primary"
                         onClick={onWriteStart}
-                        disabled={writing || writingSucceed || writingFail}
+                        disabled={
+                            writing ||
+                            writingSucceed ||
+                            writingFail ||
+                            hasChipTarget
+                        }
                     >
                         Write
                     </DialogButton>
@@ -247,6 +283,24 @@ Are you sure you want to continue?`,
                     <strong>Firmware:</strong>
                     <span>{` ${mcubootFwPath || zipFilePath}`}</span>
                 </div>
+
+                {mustHaveChipTarget && (
+                    <div className="tw-flex tw-flex-col tw-gap-2">
+                        <strong>Target:</strong>
+                        <Dropdown
+                            items={targetDropdownItems}
+                            onSelect={(item: DropdownItem) => {
+                                setChosenTarget(item.value);
+                            }}
+                            selectedItem={getSelectedDropdownItem(
+                                targetDropdownItems,
+                                chosenTarget,
+                            )}
+                            disabled={writingHasStarted}
+                        />
+                    </div>
+                )}
+
                 {writing && (
                     <div className="tw-flex tw-flex-col tw-gap-2">
                         <div>