+++ src/reducers/modemReducer.ts
@@ -6,14 +6,17 @@
 
 import { createSlice, type PayloadAction } from '@reduxjs/toolkit';
 
+import { FirmwareType } from '../util/firmwareType';
 import type { RootState } from './types';
 
 export interface ModemState {
     isVisible: boolean;
+    firmwareType: FirmwareType;
 }
 
 const initialState: ModemState = {
     isVisible: false,
+    firmwareType: FirmwareType.MODEM,
 };
 
 const modemSlice = createSlice({
@@ -23,12 +26,18 @@ const modemSlice = createSlice({
         setShowModemProgrammingDialog(state, action: PayloadAction<boolean>) {
             state.isVisible = action.payload;
         },
+        setFirmwareType(state, action: PayloadAction<FirmwareType>) {
+            state.firmwareType = action.payload;
+        },
     },
 });
 
 export const getShowModemProgrammingDialog = (state: RootState) =>
     state.app.modem.isVisible;
 
+export const getFirmwareType = (state: RootState) =>
+    state.app.modem.firmwareType;
+
 export default modemSlice.reducer;
-const { setShowModemProgrammingDialog } = modemSlice.actions;
-export { setShowModemProgrammingDialog };
+const { setShowModemProgrammingDialog, setFirmwareType } = modemSlice.actions;
+export { setShowModemProgrammingDialog, setFirmwareType };